fixed delete like

+++ app/api/like_routes.py
@@ -1,4 +1,4 @@
-from flask import Blueprint
+from flask import Blueprint, jsonify
 from app.models import db, Like, User, Post
 from flask_login import login_required, current_user
 
@@ -10,7 +10,7 @@ like_routes = Blueprint('likes', __name__)
 @like_routes.route('/posts/<int:post_id>/likes')
 @login_required
 def get_likes_by_post(post_id):
-    likes = Like.query.filter(Like.post_id == post_id)
+    likes = Like.query.filter(Like.post_id == int(post_id))
     return {'likes': [like.to_dict() for like in likes]}
 
 
@@ -20,7 +20,7 @@ def get_likes_by_post(post_id):
 def like_a_post(post_id):
     curr_user = current_user.get_id()
     new_like = Like(user_id=curr_user,
-                    post_id=post_id
+                    post_id=int(post_id)
                     )
     db.session.add(new_like)
     db.session.commit()
@@ -31,14 +31,17 @@ def like_a_post(post_id):
 @like_routes.route('/posts/<int:post_id>', methods=['DELETE'])
 @login_required
 def delete_a_like(post_id):
-    curr_user = current_user.get_id()
     post = Post.query.get(post_id)
 
     if post:
-        like = Like.query.filter(Like.user_id == curr_user,
-                                 Like.post_id == post_id)
-        like.delete()
-        db.session.commit()
-        return {"message": "Successfully deleted", "status_code": "200"}
+        like = Like.query.filter(Like.user_id == int(current_user.id),
+                                 Like.post_id == int(post_id))
+        exist = [a.to_dict() for a in like]
+        if exist:
+            like.delete()
+            db.session.commit()
+            return jsonify({"message": "Successfully deleted", "status_code": "200"})
+        else:
+            return jsonify({'message': "Like doesn't exist", 'status_code': '404'})
     else:
-        return {"message": "Post couldn't be found", "status_code": "404"}
+        return jsonify({"message": "Post couldn't be found", "status_code": "404"})
